Split apart the two generators
